requirements.txt
-e .
pandas
scikit-learn
torch
torchvision
torchaudio
imblearn
PyYAML
mlflow
flask